Extract connection
