+++ src/layouts/app/hass-base-mixin.js
@@ -4,4 +4,8 @@ export default superClass => class extends superClass {
   hassConnected() {}
   hassReconnected() {}
   panelUrlChanged(newPanelUrl) {}
+
+  _updateHass(obj) {
+    this.hass = Object.assign({}, this.hass, obj);
+  }
 };
