+++ src/layouts/app/home-assistant.js
@@ -5,22 +5,9 @@ import { html } from '@polymer/polymer/lib/utils/html-tag.js';
 import { PolymerElement } from '@polymer/polymer/polymer-element.js';
 import { afterNextRender } from '@polymer/polymer/lib/utils/render-status.js';
 
-import {
-  ERR_INVALID_AUTH,
-  subscribeEntities,
-  subscribeConfig,
-} from 'home-assistant-js-websocket';
-
-import LocalizeMixin from '../../mixins/localize-mixin.js';
-import translationMetadata from '../../../build-translations/translationMetadata.json';
 import '../../layouts/home-assistant-main.js';
 import '../../resources/ha-style.js';
-import { getState } from '../../util/ha-pref-storage.js';
-import { getActiveTranslation } from '../../util/hass-translation.js';
-import hassCallApi from '../../util/hass-call-api.js';
 import registerServiceWorker from '../../util/register-service-worker.js';
-
-import computeStateName from '../../common/entity/compute_state_name.js';
 
 import HassBaseMixin from './hass-base-mixin.js';
 import AuthMixin from './auth-mixin.js';
@@ -29,13 +16,14 @@ import ThemesMixin from './themes-mixin.js';
 import MoreInfoMixin from './more-info-mixin.js';
 import SidebarMixin from './sidebar-mixin.js';
 import DialogManagerMixin from './dialog-manager-mixin.js';
+import ConnectionMixin from './connection-mixin.js';
 
 import(/* webpackChunkName: "login-form" */ '../../layouts/login-form.js');
 
 const ext = (baseClass, mixins) => mixins.reduceRight((base, mixin) => mixin(base), baseClass);
 
 class HomeAssistant extends ext(PolymerElement, [
-  LocalizeMixin,
+  ConnectionMixin,
   DialogManagerMixin,
   AuthMixin,
   ThemesMixin,
@@ -63,7 +51,7 @@ class HomeAssistant extends ext(PolymerElement, [
     <template is="dom-if" if="[[!showMain]]" restamp>
       <login-form
         hass="[[hass]]"
-        connection-promise="{{connectionPromise}}"
+        connection-promise="[[connectionPromise]]"
         show-loading="[[computeShowLoading(connectionPromise, hass)]]"
       ></login-form>
     </template>
@@ -74,13 +62,7 @@ class HomeAssistant extends ext(PolymerElement, [
     return {
       connectionPromise: {
         type: Object,
-        value: window.hassConnection || null,
-        observer: 'handleConnectionPromise',
-      },
-      connection: {
-        type: Object,
         value: null,
-        observer: 'connectionChanged',
       },
       hass: {
         type: Object,
@@ -98,11 +80,6 @@ class HomeAssistant extends ext(PolymerElement, [
         observer: 'panelUrlChanged',
       },
     };
-  }
-
-  constructor() {
-    super();
-    this.unsubFuncs = [];
   }
 
   ready() {
@@ -124,175 +101,6 @@ class HomeAssistant extends ext(PolymerElement, [
       || (hass && hass.connection && (!hass.states || !hass.config)));
   }
 
-  connectionChanged(conn, oldConn) {
-    if (oldConn) {
-      while (this.unsubFuncs.length) {
-        this.unsubFuncs.pop()();
-      }
-    }
-    if (!conn) {
-      this._updateHass({
-        connection: null,
-        connected: false,
-        states: null,
-        config: null,
-        themes: null,
-        dockedSidebar: false,
-        moreInfoEntityId: null,
-        callService: null,
-        callApi: null,
-        sendWS: null,
-        callWS: null,
-        user: null,
-      });
-      return;
-    }
-    var notifications = this.$.notifications;
-    this.hass = Object.assign({
-      connection: conn,
-      connected: true,
-      states: null,
-      config: null,
-      themes: null,
-      panels: null,
-      panelUrl: this.panelUrl,
-
-      language: getActiveTranslation(),
-      // If resources are already loaded, don't discard them
-      resources: (this.hass && this.hass.resources) || null,
-
-      translationMetadata: translationMetadata,
-      dockedSidebar: false,
-      moreInfoEntityId: null,
-      callService: async (domain, service, serviceData = {}) => {
-        try {
-          await conn.callService(domain, service, serviceData);
-
-          let message;
-          let name;
-          if (serviceData.entity_id && this.hass.states &&
-            this.hass.states[serviceData.entity_id]) {
-            name = computeStateName(this.hass.states[serviceData.entity_id]);
-          }
-          if (service === 'turn_on' && serviceData.entity_id) {
-            message = this.localize(
-              'ui.notification_toast.entity_turned_on',
-              'entity', name || serviceData.entity_id
-            );
-          } else if (service === 'turn_off' && serviceData.entity_id) {
-            message = this.localize(
-              'ui.notification_toast.entity_turned_off',
-              'entity', name || serviceData.entity_id
-            );
-          } else {
-            message = this.localize(
-              'ui.notification_toast.service_called',
-              'service', `${domain}/${service}`
-            );
-          }
-          notifications.showNotification(message);
-        } catch (err) {
-          const msg = this.localize(
-            'ui.notification_toast.service_call_failed',
-            'service', `${domain}/${service}`
-          );
-          notifications.showNotification(msg);
-          throw err;
-        }
-      },
-      callApi: async (method, path, parameters) => {
-        const host = window.location.protocol + '//' + window.location.host;
-        const auth = conn.options;
-        try {
-          // Refresh token if it will expire in 30 seconds
-          if (auth.accessToken && Date.now() + 30000 > auth.expires) {
-            const accessToken = await window.refreshToken();
-            conn.options.accessToken = accessToken.access_token;
-            conn.options.expires = accessToken.expires;
-          }
-          return await hassCallApi(host, auth, method, path, parameters);
-        } catch (err) {
-          if (!err || err.status_code !== 401 || !auth.accessToken) throw err;
-
-          // If we connect with access token and get 401, refresh token and try again
-          const accessToken = await window.refreshToken();
-          conn.options.accessToken = accessToken.access_token;
-          conn.options.expires = accessToken.expires;
-          return await hassCallApi(host, auth, method, path, parameters);
-        }
-      },
-      // For messages that do not get a response
-      sendWS: (msg) => {
-        // eslint-disable-next-line
-        if (__DEV__) console.log('Sending', msg);
-        conn.sendMessage(msg);
-      },
-      // For messages that expect a response
-      callWS: (msg) => {
-        /* eslint-disable no-console */
-        if (__DEV__) console.log('Sending', msg);
-
-        const resp = conn.sendMessagePromise(msg);
-
-        if (__DEV__) {
-          resp.then(
-            result => console.log('Received', result),
-            err => console.log('Error', err),
-          );
-        }
-        // In the future we'll do this as a breaking change
-        // inside home-assistant-js-websocket
-        return resp.then(result => result.result);
-      },
-    }, getState());
-
-    this.hassConnected();
-  }
-
-  hassConnected() {
-    super.hassConnected();
-
-    const conn = this.hass.connection;
-
-    const reconnected = () => this.hassReconnected();
-    const disconnected = () => this._updateHass({ connected: false });
-    const reconnectError = async (_conn, err) => {
-      if (err !== ERR_INVALID_AUTH) return;
-      disconnected();
-      while (this.unsubFuncs.length) {
-        this.unsubFuncs.pop()();
-      }
-      const accessToken = await window.refreshToken();
-      this.handleConnectionPromise(window.createHassConnection(null, accessToken));
-    };
-
-    conn.addEventListener('ready', reconnected);
-    conn.addEventListener('disconnected', disconnected);
-    // If we reconnect after losing connection and access token is no longer
-    // valid.
-    conn.addEventListener('reconnect-error', reconnectError);
-
-    this.unsubFuncs.push(() => {
-      conn.removeEventListener('ready', reconnected);
-      conn.removeEventListener('disconnected', disconnected);
-      conn.removeEventListener('reconnect-error', reconnectError);
-    });
-
-    subscribeEntities(conn, states => this._updateHass({ states }))
-      .then(unsub => this.unsubFuncs.push(unsub));
-
-    subscribeConfig(conn, config => this._updateHass({ config }))
-      .then(unsub => this.unsubFuncs.push(unsub));
-
-    this._loadPanels();
-  }
-
-  hassReconnected() {
-    super.hassReconnected();
-    this._updateHass({ connected: true });
-    this._loadPanels();
-  }
-
   computePanelUrl(routeData) {
     return (routeData && routeData.panel) || 'states';
   }
@@ -302,29 +110,18 @@ class HomeAssistant extends ext(PolymerElement, [
     this._updateHass({ panelUrl: newPanelUrl });
   }
 
-  async handleConnectionPromise(prom) {
-    if (!prom) return;
+  // async handleConnectionPromise(prom) {
+  //   if (!prom) return;
 
-    try {
-      this.connection = await prom;
-    } catch (err) {
-      this.connectionPromise = null;
-    }
-  }
+  //   try {
+  //     this.connection = await prom;
+  //   } catch (err) {
+  //     this.connectionPromise = null;
+  //   }
+  // }
 
   handleNotification(ev) {
     this.$.notifications.showNotification(ev.detail.message);
-  }
-
-  async _loadPanels() {
-    const panels = await this.hass.callWS({
-      type: 'get_panels'
-    });
-    this._updateHass({ panels });
-  }
-
-  _updateHass(obj) {
-    this.hass = Object.assign({}, this.hass, obj);
   }
 }
